// Countries-SPA/src/app/models/country.ts
export class Country {
    name: string;
    code: string;
    capitalName: string;
    area: number;
    population: number;
    regionName: string;

    // constructor(name: string, code: string, capitalName: string, area: number, population: number, regionName: string) {
    //     this.name = name;
    //     this.code = code;
    //     this.capitalName = capitalName;
    //     this.area = area;
    //     this.population = population;
    //     this.regionName = regionName;
    // }
}